+++ apps/templates/index.html
@@ -1,7 +1,7 @@
 {% extends "layout.html" %}
 
 {% block content %}
-<script type="module" src="{{ url_for('js.static', filename='index.js') }}"></script>
+<script type="module" src="{{ url_for('js.static', filename='bundle.js') }}"></script>
 
 <section class="hero is-info is-medium is-bold">
     <div class="hero-body">
