+++ apps/templates/layout.html
@@ -2,9 +2,9 @@
 <html>
 <head>
 <title>YourApart</title>
-<link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='style.css') }}" />
-<link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='bulma.css') }}" />
-<script src="{{ url_for('js.static', filename='axios.js')}}"></script>
+<!-- <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='style.css') }}" />
+<link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='bulma.css') }}" /> -->
+<!-- <script src="{{ url_for('js.static', filename='axios.js')}}"></script> -->
 <meta name="viewport" content="width=device-width,initial-scale=1">
 </head>
 <body>
