+++ front/css/style.scss
@@ -1,3 +1,5 @@
+@import "~bulma";
+
 h1 {
     color: blue;
 }
