+++ front/entry.js
@@ -1,0 +1,6 @@
+// aysnc await 
+// import '@babel/polyfill';
+
+import "./css/style.scss"
+import main from "./js/index.js"
+main()
