+++ front/js/components/design.js
@@ -1,0 +1,13 @@
+import {h} from "hyperapp"
+
+export const Container = (contents) => {
+    return h("div", {}, contents)
+}
+
+export const Section = (content) => {
+    return (
+        h("section", {class: "section"}, [
+            h("div", {class: "container"}, [content])
+        ])
+    )
+}
