+++ front/js/index.js
@@ -1,0 +1,5 @@
+import index from "./pages/index.js"
+
+export default function(){
+    index()
+}
