+++ package.json
@@ -1,0 +1,34 @@
+{
+  "name": "evalhome",
+  "version": "1.0.0",
+  "description": "下宿の評価をするアプリ",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/face9363/evalhome.git"
+  },
+  "keywords": [
+    "home"
+  ],
+  "author": "face9363",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/face9363/evalhome/issues"
+  },
+  "homepage": "https://github.com/face9363/evalhome#readme",
+  "dependencies": {
+    "axios": "^0.19.2",
+    "bulma": "^0.8.0",
+    "css-loader": "^3.4.2",
+    "hyperapp": "^2.0.4",
+    "node-sass": "^4.13.1",
+    "sass-loader": "^8.0.2",
+    "style-loader": "^1.1.3",
+    "webpack": "^4.41.6",
+    "webpack-cli": "^3.3.11"
+  },
+  "devDependencies": {}
+}
